+++ README.md
@@ -3,3 +3,4 @@
 - the search filter:i got the filter to identify items with the same string as the input.value but i dont know how to display the search result
 - the items in the cart : i think this is because the items in the cart are virtual dom elements, i am having issues selecting the particular item...any selection i make selects all the elements above the present selection.... still figuring it out
 - the cart total
+hi
