+++ index.html
@@ -20,6 +20,7 @@
   <title>E-Commerce</title>
   <link rel="stylesheet" href="styles.css">
   <script defer src="index.js"></script>
+  <script defer type="module" src="list_items_file.js"></script>
 </head>
 <body>
    <main>
