--- list_items_file
@@ -1,1 +1,0 @@
-
