+++ list_items_file.js
@@ -1,0 +1,272 @@
+//objects for the cart items
+const computers=[
+   {
+     id:1,
+     itemName:'Hp pavilion 15',
+     itemPrice:'499.99',
+     description:''
+   }
+   ,{
+    id:2,
+    itemName:'Lenovo Ideapad L3 ',
+    itemPrice:'399.99',
+    description:''
+  }
+,{
+  id:3,
+  itemName:'Asus X515JA-BR068T ',
+  itemPrice:'359.99',
+  description:''
+},{
+  id:4,
+  itemName:'HP Envy',
+  itemPrice:'699.99',
+  description:''
+},{
+  id:5,
+  itemName:'Asus VivoBook  ',
+  itemPrice:'549.99 ',
+  description:''
+},{
+  id:6,
+  itemName:'Zinox Ultrabook 14 Inch ',
+  itemPrice:'149.99 ',
+  description:''
+},{
+  id:7,
+  itemName:'Lenovo 100E Intel Celeron N4000 ',
+  itemPrice:'299.99 ',
+  description:''
+},{
+  id:8,
+  itemName:'HP Stream 11 Intel Celeron',
+  itemPrice:'749.99',
+  description:''
+},{
+  id:9,
+  itemName:'Dell Vostro 3500 ',
+  itemPrice:'159.99 ',
+  description:''
+},{
+  id:10,
+  itemName:'HP Omen Gaming 15.6" ',
+  itemPrice:'999.99 ',
+  description:''
+},
+// ]
+// const Fashion=[
+  {
+    id:11,
+    itemName:"Addidas men's slides",
+    itemPrice:'39.99',
+    description:''
+  }
+  ,{
+   id:12,
+   itemName:'vintage shirt for men',
+   itemPrice:'9.99',
+   description:''
+ }
+,{
+ id:13,
+ itemName:'ladies high heel slippers ',
+ itemPrice:'14.99',
+ description:''
+},{
+ id:14,
+ itemName:'niker shorts set',
+ itemPrice:'4.99',
+ description:''
+},{
+ id:15,
+ itemName:' women hand bag  ',
+ itemPrice:'49.99 ',
+ description:''
+},{
+ id:16,
+ itemName:'ray ban sun glasses ',
+ itemPrice:'149.99 ',
+ description:''
+},{
+ id:17,
+ itemName:'ankara fabrics 6 yards ',
+ itemPrice:'2.99 ',
+ description:''
+},{
+ id:18,
+ itemName:'D&G loafers',
+ itemPrice:'749.99',
+ description:''
+},{
+ id:19,
+ itemName:'silver necklace ',
+ itemPrice:'9.99 ',
+ description:''
+},{
+  id:20,
+  itemName:'Air max shoes ',
+  itemPrice:'999.99 ',
+  description:''
+ },{
+ id:21,
+ itemName:'samsung galaxy A12',
+ itemPrice:'649.99 ',
+ description:''
+},
+{
+  id:22,
+  itemName:"tecno spark go",
+  itemPrice:'539.99',
+  description:''
+}
+,{
+ id:23,
+ itemName:'xaomi redmi note 10',
+ itemPrice:'749.99',
+ description:''
+}
+,{
+id:24,
+itemName:'infinix smart HD',
+itemPrice:'214.99',
+description:''
+},{
+id:25,
+itemName:'vivo Y20',
+itemPrice:'214.99',
+description:''
+},{
+id:26,
+itemName:'Iphone 13 pro max ',
+itemPrice:'1249.99 ',
+description:''
+},{
+id:27,
+itemName:'zinox zpad X8 ',
+itemPrice:'149.99 ',
+description:''
+},{
+id:28,
+itemName:'wireless intercom ',
+itemPrice:'452.99 ',
+description:''
+},{
+id:29,
+itemName:'samsung galaxy A1',
+itemPrice:'749.99',
+description:''
+},{
+id:30,
+itemName:'silver necklace ',
+itemPrice:'9.99 ',
+description:''
+},
+{
+  id:31,
+  itemName:'samsung galaxy A12',
+  itemPrice:'649.99 ',
+  description:''
+ },
+ {
+   id:32,
+   itemName:"tecno spark go",
+   itemPrice:'539.99',
+   description:''
+ }
+ ,{
+  id:33,
+  itemName:'xaomi redmi note 10',
+  itemPrice:'749.99',
+  description:''
+ }
+ ,{
+ id:34,
+ itemName:'infinix smart HD',
+ itemPrice:'214.99',
+ description:''
+ },{
+ id:35,
+ itemName:'vivo Y20',
+ itemPrice:'214.99',
+ description:''
+ },{
+ id:36,
+ itemName:'Iphone 13 pro max ',
+ itemPrice:'1249.99 ',
+ description:''
+ },{
+ id:37,
+ itemName:'zinox zpad X8 ',
+ itemPrice:'149.99 ',
+ description:''
+ },{
+ id:38,
+ itemName:'wireless intercom ',
+ itemPrice:'452.99 ',
+ description:''
+ },{
+ id:39,
+ itemName:'samsung galaxy A1',
+ itemPrice:'749.99',
+ description:''
+ },{
+ id:40,
+ itemName:'silver necklace ',
+ itemPrice:'9.99 ',
+ description:''
+ },{
+ id:41,
+ itemName:'Air max shoes ',
+ itemPrice:'999.99 ',
+ description:''
+ },{
+ id:42,
+ itemName:' ',
+ itemPrice:' ',
+ description:''
+ },{
+id:43,
+itemName:'Air max shoes ',
+itemPrice:'999.99 ',
+description:''
+},{
+id:44,
+itemName:' ',
+itemPrice:' ',
+description:''
+},
+{
+  id:45,
+  itemName:'Air max shoes ',
+  itemPrice:'999.99 ',
+  description:''
+ },
+ {
+    id:46,
+    itemName:'Air max shoes ',
+    itemPrice:'999.99 ',
+    description:''
+   },{
+    id:47,
+    itemName:'Air max shoes ',
+    itemPrice:'999.99 ',
+    description:''
+   },{
+    id:48,
+    itemName:'Air max shoes ',
+    itemPrice:'999.99 ',
+    description:''
+   },{
+    id:49,
+    itemName:'Air max shoes ',
+    itemPrice:'999.99 ',
+    description:''
+   },{
+    id:50,
+    itemName:'Air max shoes ',
+    itemPrice:'999.99 ',
+    description:''
+   }
+];
+
+export {computers};
